Part X.

+++ src/server/game/AuctionHouse/AuctionHouseMgr.cpp
@@ -707,13 +707,13 @@ bool AuctionEntry::BuildAuctionInfo(WorldPacket& data) const
     data << uint32(item->GetCount());                               // item->count
     data << uint32(item->GetSpellCharges());                        // item->charge FFFFFFF
     data << uint32(0);                                              // Unknown
-    data << uint64(owner);                                          // Auction->owner
+    data << owner;                                                  // Auction->owner
     data << uint32(startbid);                                       // Auction->startbid (not sure if useful)
     data << uint32(bid ? GetAuctionOutBid() : 0);
     // Minimal outbid
     data << uint32(buyout);                                         // Auction->buyout
-    data << uint32((expire_time - time(nullptr)) * IN_MILLISECONDS);   // time left
-    data << uint64(bidder);                                         // auction->bidder current
+    data << uint32((expire_time - time(nullptr)) * IN_MILLISECONDS); // time left
+    data << bidder;                                                 // auction->bidder current
     data << uint32(bid);                                            // current bid
     return true;
 }
